feat: update new release workflow
