+++ release_notes/input_repos.json
@@ -1,0 +1,5 @@
+[
+    "ascend-io/ascend-ui",
+    "ascend-io/ascend-core",
+    "ascend-io/ascend-backend"
+]
